event-protocol: Extract validateEvent function
