+++ event-protocol/validate.js
@@ -10,37 +10,27 @@
  * The script exits with 1 if one or more events are invalid.
  */
 
-const inspect = require('util').inspect
 const path = require('path')
 const es = require('event-stream')
 const fs = require('mz/fs')
-const Ajv = require('ajv');
 const RED = '\033[0;31m'
 const GREEN = '\033[0;32m'
 const NC = '\033[0m'
-const SCHEMA_BASE_URL = 'https://raw.github.com/cucumber/cucumber/master/event-protocol/schemas/'
+const validateEvent = require('./lib/validateEvent')
 
-const schemaDir = __dirname + '/schemas'
-const schemaFiles = fs.readdirSync(schemaDir)
-  .map(name => path.resolve(`${schemaDir}/${name}`))
-
-var ajv = Ajv({
-  allErrors: true,
-  schemas: schemaFiles.map(require)
-})
-
-const validateEvents = (schemas) => {
+const validateEvents = () => {
   let validationError = false
   return new Promise((resolve, reject) => {
     process.stdin
       .pipe(es.split())
       .pipe(es.map((json, cb) => json ? cb(null, JSON.parse(json)) : cb()))
       .pipe(es.map((event, cb) => {
-        const validate = schemas.getSchema(`${SCHEMA_BASE_URL}${event.type}.json#`)
-        if (!validate) return cb(null, [`No schema for ${inspect(event)}`, event])
-        const valid = validate(event)
-        if (!valid) return cb(null, [JSON.stringify(validate.errors), event])
-        cb(null, [null, event])
+        try {
+          validateEvent(event)
+          cb(null, [null, event])
+        } catch(err) {
+          cb(null, [err.message, event])
+        }
       }))
       .pipe(es.map(([errorMessage, event], cb) => {
         if (errorMessage) {
@@ -59,7 +49,7 @@ const validateEvents = (schemas) => {
   })
 }
 
-validateEvents(ajv)
+validateEvents()
   .then(validationError => process.exit(validationError ? 1 : 0))
   .catch(err => {
     console.error(`${RED}${err}${NC}`)
